renamed templates to patterns
