+++ edxpipelines/patterns/__init__.py
@@ -1,0 +1,4 @@
+from edxpipelines.patterns import jobs
+from edxpipelines.patterns import piplines
+from edxpipelines.patterns import stages
+from edxpipelines.patterns import tasks
