--- edxpipelines/templates/__init__.py
@@ -1,4 +1,0 @@
-from edxpipelines.templates import jobs
-from edxpipelines.templates import piplines
-from edxpipelines.templates import stages
-from edxpipelines.templates import tasks
